chore: improve tsdoc
